Mixin only for tabs.
Required so I can easily overwrite these methods in Pro.
Damn you, Ruby inheritance!
