+++ app/mixins/tabs.rb
@@ -1,10 +1,38 @@
 module Tabs
-  include TabsHelper
-
   protected
 
   def verify_tab
     params.delete(:tab) if params[:tab] && !display_tab?(params[:tab])
     params[:tab] ||= tabs.first
+  end
+
+  def tabs
+    @tabs ||= %w(repos profile).select { |tab| display_tab?(tab) }
+  end
+
+  def current_tab
+    params[:tab]
+  end
+
+  def display_tab?(tab)
+    tab != 'profile' || owner == current_user
+  end
+
+  def owner
+    @owner ||= params[:owner_name] ? owners.detect { |owner| owner.login == params[:owner_name] } : current_user
+  end
+
+  def owners
+    @owners ||= [current_user] + Organization.where(:login => owner_names)
+  end
+
+  def repository_counts
+    @repository_counts ||= Repository.counts_by_owner_names(owner_names)
+  end
+
+  def self.included(base)
+    %w{owner tab_path current_tab repository_counts tabs owners}.each do |helper|
+      base.helper_method helper if base.respond_to?(:helper_method)
+    end
   end
 end
